Updated to use portals
