+++ interface/app/$libraryId/overview.tsx
@@ -17,7 +17,7 @@ import { Statistics, useLibraryContext, useLibraryQuery } from '@sd/client';
 import { Card, ScreenHeading } from '@sd/ui';
 import useCounter from '~/hooks/useCounter';
 import { usePlatform } from '~/util/Platform';
-import { useToolBar } from './TopBar/useToolBar';
+import TopBarChildren from './TopBar/TopBarChildren';
 
 interface StatItemProps {
 	title: string;
@@ -57,9 +57,6 @@ const StatItem = (props: StatItemProps) => {
 		end: +size.value,
 		duration: overviewMounted ? 0 : 1,
 		saveState: false
-	});
-	useToolBar({
-		options: [[]]
 	});
 
 	return (
@@ -104,44 +101,47 @@ export const Component = () => {
 	overviewMounted = true;
 
 	return (
-		<div className="flex h-screen w-full flex-col">
-			<ScreenHeading>Overview</ScreenHeading>
-			{/* STAT HEADER */}
-			<div className="flex w-full">
-				{/* STAT CONTAINER */}
-				<div className="-mb-1 flex h-20 overflow-hidden">
-					{Object.entries(stats?.data || []).map(([key, value]) => {
-						if (!displayableStatItems.includes(key)) return null;
-						return (
-							<StatItem
-								key={`${library.uuid} ${key}`}
-								title={StatItemNames[key as keyof Statistics]!}
-								bytes={BigInt(value)}
-								isLoading={platform.demoMode ? false : stats.isLoading}
-							/>
-						);
-					})}
+		<>
+			<TopBarChildren />
+			<div className="flex h-screen w-full flex-col">
+				<ScreenHeading>Overview</ScreenHeading>
+				{/* STAT HEADER */}
+				<div className="flex w-full">
+					{/* STAT CONTAINER */}
+					<div className="-mb-1 flex h-20 overflow-hidden">
+						{Object.entries(stats?.data || []).map(([key, value]) => {
+							if (!displayableStatItems.includes(key)) return null;
+							return (
+								<StatItem
+									key={`${library.uuid} ${key}`}
+									title={StatItemNames[key as keyof Statistics]!}
+									bytes={BigInt(value)}
+									isLoading={platform.demoMode ? false : stats.isLoading}
+								/>
+							);
+						})}
+					</div>
+					<div className="grow" />
 				</div>
-				<div className="grow" />
+				<div className="mt-4 grid grid-cols-5 gap-3 pb-4">
+					<CategoryButton icon={Heart} category="Favorites" />
+					<CategoryButton icon={FileText} category="Documents" />
+					<CategoryButton icon={Camera} category="Movies" />
+					<CategoryButton icon={FrameCorners} category="Screenshots" />
+					<CategoryButton icon={AppWindow} category="Applications" />
+					<CategoryButton icon={Wrench} category="Projects" />
+					<CategoryButton icon={CloudArrowDown} category="Downloads" />
+					<CategoryButton icon={MusicNote} category="Music" />
+					<CategoryButton icon={Image} category="Albums" />
+					<CategoryButton icon={Heart} category="Favorites" />
+				</div>
+				<Card className="text-ink-dull">
+					<b>Note: </b>&nbsp; This is a pre-alpha build of Spacedrive, many features are
+					yet to be functional.
+				</Card>
+				<div className="flex h-4 w-full shrink-0" />
 			</div>
-			<div className="mt-4 grid grid-cols-5 gap-3 pb-4">
-				<CategoryButton icon={Heart} category="Favorites" />
-				<CategoryButton icon={FileText} category="Documents" />
-				<CategoryButton icon={Camera} category="Movies" />
-				<CategoryButton icon={FrameCorners} category="Screenshots" />
-				<CategoryButton icon={AppWindow} category="Applications" />
-				<CategoryButton icon={Wrench} category="Projects" />
-				<CategoryButton icon={CloudArrowDown} category="Downloads" />
-				<CategoryButton icon={MusicNote} category="Music" />
-				<CategoryButton icon={Image} category="Albums" />
-				<CategoryButton icon={Heart} category="Favorites" />
-			</div>
-			<Card className="text-ink-dull">
-				<b>Note: </b>&nbsp; This is a pre-alpha build of Spacedrive, many features are yet
-				to be functional.
-			</Card>
-			<div className="flex h-4 w-full shrink-0" />
-		</div>
+		</>
 	);
 };
