+++ interface/app/$libraryId/spacedrop.tsx
@@ -6,7 +6,7 @@ import { Button, ScreenHeading, Select, SelectOption, forms, tw } from '@sd/ui';
 import { PeerMetadata, useBridgeMutation, useBridgeSubscription } from '~/../packages/client/src';
 import { SubtleButton, SubtleButtonContainer } from '~/components/SubtleButton';
 import { OperatingSystem } from '~/util/Platform';
-import { useToolBar } from './TopBar/useToolBar';
+import TopBarChildren from './TopBar/TopBarChildren';
 import classes from './spacedrop.module.scss';
 
 const { Form, Input, useZodForm, z } = forms;
@@ -163,11 +163,9 @@ function TemporarySpacedropDemo() {
 }
 
 export const Component = () => {
-	useToolBar({
-		options: [[]]
-	});
 	return (
 		<>
+			<TopBarChildren toolOptions={[[]]} />
 			<TemporarySpacedropDemo />
 			<div className={classes.honeycombOuter}>
 				<div className={clsx(classes.honeycombContainer, 'mt-8')}>
