+++ interface/index.tsx
@@ -10,7 +10,6 @@ import { ErrorBoundary } from 'react-error-boundary';
 import { RouterProvider, RouterProviderProps } from 'react-router-dom';
 import { useDebugState } from '@sd/client';
 import ErrorFallback from './ErrorFallback';
-import ToolBarProvider from './app/$libraryId/TopBar/ToolBarProvider';
 
 export * from './util/keybind';
 export * from './util/Platform';
@@ -48,9 +47,7 @@ export const SpacedriveInterface = (props: { router: RouterProviderProps['router
 	return (
 		<ErrorBoundary FallbackComponent={ErrorFallback}>
 			<Devtools />
-			<ToolBarProvider>
-				<RouterProvider router={props.router} />
-			</ToolBarProvider>
+			<RouterProvider router={props.router} />
 		</ErrorBoundary>
 	);
 };
